// superset-frontend/plugins/superset-plugin-chart-hazard-risk-outlook/src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { DataRecord, QueryFormColumn, QueryFormData } from '@superset-ui/core';

export interface SupersetPluginChartHazardRiskOutlookStylesProps {
  height: number;
  width: number;
}

interface SupersetPluginChartHazardRiskOutlookCustomizeProps {
  forecastStartMilliseconds: number;
  headerText: string;
  hazardColumn: string;
  severityColumn: string;
  validFromColumn: string;
  validToColumn: string;
}

export type SupersetPluginChartHazardRiskOutlookQueryFormData =
  QueryFormData & {
    hazard_column?: QueryFormColumn;
    severity_column?: QueryFormColumn;
    valid_from_column?: QueryFormColumn;
    valid_to_column?: QueryFormColumn;
    hazardColumn?: QueryFormColumn;
    severityColumn?: QueryFormColumn;
    validFromColumn?: QueryFormColumn;
    validToColumn?: QueryFormColumn;
  };

export type HazardRiskOutlookDataRecord = DataRecord & {
  duration_hours: number | null;
};

export type SupersetPluginChartHazardRiskOutlookProps =
  SupersetPluginChartHazardRiskOutlookStylesProps &
    SupersetPluginChartHazardRiskOutlookCustomizeProps & {
      data: HazardRiskOutlookDataRecord[];
    };